refactor: move test utils into test folder
